add precedence_negative_literal lint
